Updated cranelift-jit for reduced types

+++ interpreter/src/runtime/interpreter/expressions/lists.rs
@@ -98,7 +98,7 @@ impl InterpreterEnvironment {
                                 RefMutability::Value,
                                 None,
                             )],
-                            vec![(Node::new(NodeType::IntLiteral(i as i128), range_node.line, range_node.col), None)],
+                            vec![(Node::new(NodeType::IntLiteral(i), range_node.line, range_node.col), None)],
                         )?;
                         if self.handle_conditionals(&new_scope, conditionals.clone())? {
                             result.push(
